Zmiana typu cen: float zmienione na int (ilość groszy).

+++ giepp/src/pl/pisz/airlog/giepp/data/ArchivedStock.java
@@ -5,10 +5,10 @@ public class ArchivedStock {
 
 	protected String name;
 	private String date;
-	protected float minPrice;
-	protected float maxPrice;
+	protected int minPrice;
+	protected int maxPrice;
 	
-	public ArchivedStock(String name, float maxPrice, float minPrice) {
+	public ArchivedStock(String name, int maxPrice, int minPrice) {
 		this.name = name; 
 		this.maxPrice = maxPrice; 
 		this.minPrice = minPrice; 
@@ -16,8 +16,8 @@ public class ArchivedStock {
 	public void setDate(String date) {	this.date = date; }
 	public String getName() { return name; }
 	public String getDate() { return date; }
-	public float getMinPrice() { return minPrice; }
-	public float getMaxPrice() { return maxPrice; }
+	public int getMinPrice() { return minPrice; }
+	public int getMaxPrice() { return maxPrice; }
 
 }
